requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.7
cron-descriptor==1.4.0
cycler==0.11.0
Django==4.1.7
django-bootstrap4==23.1
django-celery-beat==2.5.0
django-channels==0.7.0
django-tables2==2.5.3
django-timezone-field==5.0
djangorestframework==3.14.0
fonttools==4.39.4
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.1
httpx==0.24.0
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
kombu==5.2.4
matplotlib==3.7.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pika==1.3.2
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pyparsing==3.0.9
pyTelegramBotAPI==4.12.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.5
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.4
zipp==3.15.0